Mudado para Long o tipo do parâmetro de carregaEventoById()

+++ app/src/main/java/com/infact/nightour/controller/EventosController.java
@@ -65,7 +65,7 @@ public class EventosController {
         return cursor;
     }
 
-    public Cursor carregaEventoById(int id) {
+    public Cursor carregaEventoById(Long id) {
         String[] campos = camposTabelaEvento();
         String where = Evento.BD_ID + " = " + id;
